BayesMallows: Bayesian Preference Learning with the Mallows Rank Model.

The BayesMallows package provides functionality for fully Bayesian analysis of preference or rank data. The package implements the Bayesian Mallows model described in Vitelli et al. (2018) , which handles complete rankings, top-k rankings, ranks missing at random, and consistent pairwise preference data, as well as mixtures of rank models. Modeling of pairwise preferences containing inconsistencies, as described in Crispino et al. (2019) , is also supported. See also Sørensen et al. (2020) for an overview of the methods and a tutorial.

The documentation and examples for the following functions are likely most useful to get you started:

For analysis of rank or preference data, see compute_mallows.

For computation of multiple models with varying numbers of mixture components, see compute_mallows_mixtures.

For estimation of the partition function (normalizing constant) using either the importance sampling algorithm of Vitelli et al. (2018) or the asymptotic algorithm of Mukherjee (2016) , see estimate_partition_function.

For sequential Monte Carlo algorithms developed in Stein (2023) , see smc_mallows_new_users and smc_mallows_new_item_rank.

References

Crispino M, Arjas E, Vitelli V, Barrett N, Frigessi A (2019). “A Bayesian Mallows approach to nontransitive pair comparison data: How human are sounds?” The Annals of Applied Statistics, 13(1), 492--519. doi:10.1214/18-aoas1203 .

Mukherjee S (2016). “Estimation in exponential families on permutations.” The Annals of Statistics, 44(2), 853--875. doi:10.1214/15-aos1389 .

Sørensen Ø, Crispino M, Liu Q, Vitelli V (2020). “BayesMallows: An R Package for the Bayesian Mallows Model.” The R Journal, 12(1), 324--342. doi:10.32614/RJ-2020-026 .

Stein A (2023). Sequential Inference with the Mallows Model. Ph.D. thesis, Lancaster University. 2023-08-26.

Vitelli V, Sørensen, Crispino M, Arjas E, Frigessi A (2018). “Probabilistic Preference Learning with the Mallows Rank Model.” Journal of Machine Learning Research, 18(1), 1--49. https://jmlr.org/papers/v18/15-481.html.