Compute Posterior Intervals

Compute posterior intervals of parameters of interest.

Usage

compute_posterior_intervals(model_fit, ...)

Arguments

model_fit

A model object.

...

other arguments passed to methods.

Details

This function computes both the Highest Posterior Density Interval (HPDI), which may be discontinuous for bimodal distributions, and the central posterior interval, which is simply defined by the quantiles of the posterior distribution. The HPDI intervals are computed using the HDInterval package (Meredith and Kruschke 2018) .

References

Meredith M, Kruschke J (2018). HDInterval: Highest (Posterior) Density Intervals. R package version 0.2.0, https://CRAN.R-project.org/package=HDInterval.

See also

Examples

# The example datasets potato_visual and potato_weighing contain complete
# rankings of 20 items, by 12 assessors. We first analyse these using the Mallows
# model:
model_fit <- compute_mallows(potato_visual)

# Se the documentation to compute_mallows for how to assess the convergence of the algorithm
# Having chosen burin = 1000, we compute posterior intervals
model_fit$burnin <- 1000
# First we compute the interval for alpha
compute_posterior_intervals(model_fit, parameter = "alpha")
#>   parameter   mean median conf_level           hpdi central_interval
#> 1     alpha 10.801 10.783       95 % [9.484,12.101]   [9.468,12.089]
# We can reduce the number decimals
compute_posterior_intervals(model_fit, parameter = "alpha", decimals = 2)
#>   parameter mean median conf_level         hpdi central_interval
#> 1     alpha 10.8  10.78       95 % [9.48,12.10]     [9.47,12.09]
# By default, we get a 95 % interval. We can change that to 99 %.
compute_posterior_intervals(model_fit, parameter = "alpha", level = 0.99)
#>   parameter   mean median conf_level           hpdi central_interval
#> 1     alpha 10.801 10.783       99 % [9.075,12.437]   [9.177,12.582]
# We can also compute the posterior interval for the latent ranks rho
compute_posterior_intervals(model_fit, parameter = "rho")
#>    item parameter mean median conf_level    hpdi central_interval
#> 1    P1       rho   10     10       95 %  [9,12]           [9,12]
#> 2    P2       rho   17     17       95 % [16,18]          [16,18]
#> 3    P3       rho   19     19       95 %    [19]             [19]
#> 4    P4       rho   16     16       95 % [16,18]          [16,18]
#> 5    P5       rho   10     10       95 %  [9,12]           [9,12]
#> 6    P6       rho   15     15       95 %    [15]             [15]
#> 7    P7       rho    6      6       95 %   [5,7]            [5,7]
#> 8    P8       rho   20     20       95 %    [20]             [20]
#> 9    P9       rho    3      3       95 %     [3]              [3]
#> 10  P10       rho    4      4       95 %     [4]            [4,5]
#> 11  P11       rho   10     10       95 %  [9,12]           [8,12]
#> 12  P12       rho    1      1       95 %     [1]              [1]
#> 13  P13       rho    2      2       95 %     [2]              [2]
#> 14  P14       rho    7      7       95 %   [6,7]            [6,7]
#> 15  P15       rho   18     18       95 % [17,18]          [17,18]
#> 16  P16       rho    8      8       95 %     [8]            [8,9]
#> 17  P17       rho    5      5       95 %   [5,7]            [4,7]
#> 18  P18       rho   14     14       95 % [13,14]          [13,14]
#> 19  P19       rho   11     12       95 %  [9,12]           [9,12]
#> 20  P20       rho   13     13       95 % [13,14]          [13,14]

if (FALSE) {
  # Posterior intervals of cluster probabilities
  # We can run a mixture of Mallows models, using the n_clusters argument
  # We use the sushi example data. See the documentation of compute_mallows for a more elaborate
  # example
  model_fit <- compute_mallows(sushi_rankings, n_clusters = 5)
  # Keeping the burnin at 1000, we can compute the posterior intervals of the cluster probabilities
  compute_posterior_intervals(model_fit, burnin = 1000, parameter = "cluster_probs")
}