Compute Posterior Intervals Rho

posterior confidence intervals for rho

Usage

compute_posterior_intervals_rho(
  output,
  nmc,
  burnin,
  colnames = NULL,
  verbose = FALSE
)

Arguments

output

a subset of an SMCMallows object (though technically any matrix will do)

nmc

Number of Monte Carlo samples

burnin

A numeric value specifying the number of iterations to discard as burn-in.

colnames

colnames

verbose

if TRUE, prints the final output even if the function is assigned to an object. Defaults to FALSE.

See also

Author

Anja Stein

Examples

# Basic elements
data <- sushi_rankings[1:100, ]
n_items <- ncol(data)
leap_size <- floor(n_items / 5)
metric <- "footrule"
Time <- 20
N <- 100

# Prepare exact partition function
cardinalities <- prepare_partition_function(metric = metric,
                                            n_items = n_items)$cardinalities

# Performing SMC
smc_test <- smc_mallows_new_users(
  R_obs = data, type = "complete", n_items = n_items,
  metric = metric, leap_size = leap_size,
  N = N, Time = Time,
  cardinalities = cardinalities,
  mcmc_kernel_app = 5,
  num_new_obs = 5,
  alpha_prop_sd = 0.5,
  lambda = 0.15,
  alpha_max = 1e6
)

test_sample_rho <- smc_test$rho_samples[, , Time + 1]
compute_posterior_intervals_rho(
  output = test_sample_rho,
  nmc = N, burnin = 0,
  verbose = FALSE
)
#>                      item parameter mean median conf_level  hpdi
#> Cluster 1.Item 1   Item 1       rho    4      4       95 % [3,5]
#> Cluster 1.Item 2   Item 2       rho    4      4       95 % [2,5]
#> Cluster 1.Item 3   Item 3       rho    2      2       95 % [2,3]
#> Cluster 1.Item 4   Item 4       rho    6      6       95 % [6,8]
#> Cluster 1.Item 5   Item 5       rho    8      8       95 % [7,9]
#> Cluster 1.Item 6   Item 6       rho    4      4       95 % [2,5]
#> Cluster 1.Item 7   Item 7       rho    9      9       95 % [8,9]
#> Cluster 1.Item 8   Item 8       rho    1      1       95 %   [1]
#> Cluster 1.Item 9   Item 9       rho    7      7       95 % [6,8]
#> Cluster 1.Item 10 Item 10       rho   10     10       95 %  [10]
#>                   central_interval
#> Cluster 1.Item 1             [2,5]
#> Cluster 1.Item 2             [2,5]
#> Cluster 1.Item 3             [2,4]
#> Cluster 1.Item 4             [5,8]
#> Cluster 1.Item 5             [6,9]
#> Cluster 1.Item 6             [2,5]
#> Cluster 1.Item 7             [8,9]
#> Cluster 1.Item 8               [1]
#> Cluster 1.Item 9             [6,8]
#> Cluster 1.Item 10             [10]

compute_rho_consensus(
  output = test_sample_rho, nmc = N,
  burnin = 0, C = 1, type = "CP",
  verbose = FALSE
)
#>            ranking    item cumprob
#> cluster 1        1  Item 8    1.00
#> cluster 11       2  Item 3    0.90
#> cluster 12       3  Item 1    0.45
#> cluster 13       4  Item 6    0.57
#> cluster 14       5  Item 2    0.98
#> cluster 15       6  Item 4    0.72
#> cluster 16       7  Item 9    0.96
#> cluster 17       8  Item 5    0.91
#> cluster 18       9  Item 7    0.98
#> cluster 19      10 Item 10    1.00

test_sample_alpha <- smc_test$alpha_samples[, Time + 1]
compute_posterior_intervals_alpha(
  output = test_sample_alpha,
  nmc = N, burnin = 0, verbose = FALSE
)
#>           parameter mean median conf_level        hpdi central_interval
#> Cluster 1     alpha 1.73   1.72       95 % [1.54,2.03]      [1.49,1.99]