Expected value of metrics under a Mallows rank model

Compute the expectation of several metrics under the Mallows rank model.

Usage

expected_dist(alpha, n_items, metric)

Arguments

alpha

Non-negative scalar specifying the scale (precision) parameter in the Mallows rank model.

n_items

Integer specifying the number of items.

metric

Character string specifying the distance measure to use. Available options are "kendall", "cayley", "hamming", "ulam" for n_items<=95, "footrule" for n_items<=50 and "spearman" for n_items<=14.

Value

A scalar providing the expected value of the metric under the Mallows rank model with distance specified by the metric argument.

See also

Examples

expected_dist(1, 5, metric = "kendall")
#> [1] 4.177277
expected_dist(2, 6, metric = "cayley")
#> [1] 3.212053
expected_dist(1.5, 7, metric = "hamming")
#> [1] 5.761023
expected_dist(5, 30, "ulam")
#> [1] 21.33016
expected_dist(3.5, 45, "footrule")
#> [1] 377.5987
expected_dist(4, 10, "spearman")
#> [1] 7.220669