A synthetic 3D matrix generated using the sample_mallows function

A synthetic 3D matrix (n_users, n_items, Time) generated using the sample_mallows function. These are test datasets used to run the SMC-Mallows framework for the cases where we know all of the users in our system and their original ranking information are partial rankings. However at some point in time, we observe extra information about an existing user in the form of a rank for an item that was previously not known (NA). These datasets are very contrived as the first time step (sample_dataset[, , 1]) we observed the top m / 2 items from each user, where m is the number of items in a ranking. Then, as we increase the time, we observe the next top ranked item from one user at a time, then the next top ranked item, and so on until we have a complete dataset at sample_dataset[, , Time].

Usage

sample_dataset

Format

An object of class array of dimension 10 x 6 x 31.

References

https://github.com/anjastein/SMC-Mallows/tree/main/data

See also