Set progress report options for MCMC algorithm

Specify whether progress should be reported, and how often.

Usage

set_progress_report(verbose = FALSE, report_interval = 1000)

Arguments

verbose

Boolean specifying whether to report progress or not. Defaults to FALSE.

report_interval

Strictly positive number specifying how many iterations of MCMC should be run between each progress report. Defaults to 1000.

Value

An object of class "BayesMallowsProgressReport", to be provided in the progress_report argument to compute_mallows() and compute_mallows_mixtures().

References

There are no references for Rd macro \insertAllCites on this help page.

See also